requirements.txt
aiogram<4.0
pydantic-settings
psycopg2
huggingface_hub
onnxruntime
pillow
numpy
tqdm
torch
torchvision
pillow
tqdm
numpy